{"schema":"decodex/commit/1","summary":"Split manual and app-server tests into flat modules","authority":"manual"}
